references/agents/hello-agents/Co-creation-projects/YYHDBL-HelloCodeAgentCli/requirement.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
hello-agents[all]=0.2.7